requirements.txt
SpeechRecognition
pandas
google-generativeai
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Cors==3.0.10
sqlalchemy<2.0
psycopg2==2.9.6
Flask-Caching==1.10.1